test(functional): satisfy provider fixture size limits
